Retune dark theme surface and border tokens
Consolidates three theme guard files into one themeContract.test.ts
(dropdownSurfaces.contract.test.ts and themePairTokens.test.ts are gone,
folded in). Alongside that: restored the token-driven "paper" texture on
rich-text editors; moved the dark app-shell ground from border-100 to
border-50 so cards actually clear it; raised the dark border ramp to an
18-luma floor (border-200 sat only 8.4 off the card, a third of light's
23.7) without weakening pairs already above it; retuned alga's skeleton
shades to match; and aligned the dark header background to the sidebar
background in four pairs (slate, sky, cappuccino, vice) that had drifted
apart.

Every border in the dark theme had gone the way of the Cheshire Cat —
technically present, 8.4 luma of grin floating over the card and nothing
else. We turned the dial past 18 so the smile stays put, gave the app-
shell ground somewhere solid to stand, and stitched four wandering
header bars back onto their sidebars. 😺🌚
